backend/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
flask>=3.0.0
flask-cors>=4.0.0
pycryptodome>=3.20.0
pydantic>=2.0.0
